requirements.txt
django
gunicorn
django-heroku
django-twilio
django-solo
django-phonenumber-field
pytz
django-sendgrid-v5
